agentic-ai/langchain/requirements.txt
langchain
langchain-core
langchain-community
langchain-openai
langchain-google-genai
langgraph
fastapi
uvicorn
python-dotenv
pydantic